requirements.txt
streamlit
huggingface_hub
google-generativeai
yt-dlp